minimal fix

+++ main.py
@@ -15,7 +15,10 @@ import os
 from pathlib import Path
 
 from config import (
-    TELEGRAM_TOKEN
+    TELEGRAM_TOKEN,
+    ADMIN_ID_USER,
+    CONFUSION_MATRIX_FILE_PATH,
+    LEARNING_CURVE_FILE_PATH
 )
 from bot_logic import LunchMindBot
 
@@ -37,61 +40,84 @@ async def start_command(update: Update, context: ContextTypes.DEFAULT_TYPE):
     await update.message.reply_text(
         "Привет! Я бот для заказа обедов из ресторана. Чем могу помочь?",
         reply_markup=bot.menu_keyboard,
-        parse_mode='Markdown'
+        parse_mode='HTML'
     )
 
 
 async def help_command(update: Update, context: ContextTypes.DEFAULT_TYPE):
     """Вспомогательная функция для вывода сообщения с помощью"""
     await update.message.reply_text(
-        "Я могу помочь с меню, оформлением заказа, информацией о работе ресторана и просто поддержать беседу. Просто напишите!",
+        "Я могу помочь с меню, оформлением заказа, информацией о работе ресторана и просто поддержать беседу. Напишите мне!",
         reply_markup=bot.menu_keyboard,
-        parse_mode='Markdown'
+        parse_mode='HTML'
     )
 
 
+async def graphics_command(update: Update, context: ContextTypes.DEFAULT_TYPE):
+    """Вспомогательная функция для отправки графиков обучения"""
+    if os.path.exists(CONFUSION_MATRIX_FILE_PATH):
+        with open(CONFUSION_MATRIX_FILE_PATH, 'rb') as photo:
+            await update.message.reply_photo(
+                photo=photo,
+                caption="📊 <b>Матрица ошибок</b>\n\n<i>Это график, показывающий, как модель классифицирует интенты</i>",
+                parse_mode='HTML'
+            )
+    else:
+        await update.message.reply_text(f"❌ Файл не найден: {CONFUSION_MATRIX_FILE_PATH}")
+
+    if os.path.exists(LEARNING_CURVE_FILE_PATH):
+        with open(LEARNING_CURVE_FILE_PATH, 'rb') as photo:
+            await update.message.reply_photo(
+                photo=photo,
+                caption="📈 <b>Кривая обучения</b>\n\n<i>Показывает, как модель обучалась на разных размерах выборки</i>",
+                parse_mode='HTML'
+            )
+    else:
+        await update.message.reply_text(f"❌ Файл не найден: {LEARNING_CURVE_FILE_PATH}")
+
+
 async def menu_command(update: Update, context: ContextTypes.DEFAULT_TYPE):
-    """Функция обращается к боту вывода меню"""
+    """Функция обращается к боту для вывода меню"""
     await update.message.reply_text(
         bot.show_menu(),
         reply_markup=bot.menu_keyboard,
-        parse_mode='Markdown'
+        parse_mode='HTML'
     )
 
 
 async def cart_command(update: Update, context: ContextTypes.DEFAULT_TYPE):
     """Функция обращается к боту для вывода корзины"""
-    user_id = str(update.message.from_user.id)
+    user_id = str(update.effective_user.id)
     await update.message.reply_text(
         bot.show_cart(user_id),
         reply_markup=bot.menu_keyboard,
-        parse_mode='Markdown'
+        parse_mode='HTML'
     )
 
 
 async def clear_cart_command(update: Update, context: ContextTypes.DEFAULT_TYPE):
     """Функция обращается к боту для очистки корзины"""
-    user_id = str(update.message.from_user.id)
+    user_id = str(update.effective_user.id)
     await update.message.reply_text(
         bot.clear_cart(user_id),
         reply_markup=bot.menu_keyboard,
-        parse_mode='Markdown'
+        parse_mode='HTML'
     )
 
 
 async def complete_order_command(update: Update, context: ContextTypes.DEFAULT_TYPE):
     """Функция обращается к боту для совершения заказа"""
-    user_id = str(update.message.from_user.id)
+    user_id = str(update.effective_user.id)
     await update.message.reply_text(
         bot.complete_order(user_id),
         reply_markup=bot.menu_keyboard,
-        parse_mode='Markdown'
+        parse_mode='HTML'
     )
 
 
 async def handle_message_text(update: Update, context: ContextTypes.DEFAULT_TYPE):
     """Обрабатывает текстовые сообщения."""
-    user_id = str(update.message.from_user.id)
+    user_id = str(update.effective_user.id)
     user_name = update.effective_user.username
     text = update.message.text
 
@@ -111,13 +137,25 @@ async def handle_message_text(update: Update, context: ContextTypes.DEFAULT_TYPE
         await complete_order_command(update, context)
         logger.info(f"Пользователь ({user_id} - {user_name}): нажал кнопку \"оформить заказ\"")
 
+    elif text == "Графики обучения":
+        if user_id == ADMIN_ID_USER:
+            await graphics_command(update, context)
+        else:
+            await update.message.reply_text(
+                "<b>Приношу свои извинения!</b>\n"
+                "Этой командой могут пользоваться только <u>администраторы</u>.",
+                reply_markup=bot.menu_keyboard,
+                parse_mode='HTML'
+            )
+        logger.info(f"Пользователь ({user_id} - {user_name}): ввел команду администратора \"Графики обучения\"")
+
     else:
         response = bot.handle_message(text, user_id)
 
         await update.message.reply_text(
             response,
             reply_markup=bot.menu_keyboard,
-            parse_mode='Markdown'
+            parse_mode='HTML'
         )
 
         logger.info(f"Сообщение от пользователя ({user_id} - {user_name}): {text}")
@@ -132,20 +170,20 @@ async def handle_message_text(update: Update, context: ContextTypes.DEFAULT_TYPE
                 await update.message.reply_text(
                     sorry_message,
                     reply_markup=bot.menu_keyboard,
-                    parse_mode='Markdown'
+                    parse_mode='HTML'
                 )
         # Выдаем купон на скидку за каждые 5 покупок
         elif coupon_message:
             await update.message.reply_text(
                 coupon_message,
                 reply_markup=bot.menu_keyboard,
-                parse_mode='Markdown'
+                parse_mode='HTML'
             )
 
 
 async def handle_voice_message(update: Update, context: ContextTypes.DEFAULT_TYPE):
     """Обрабатывает голосовые сообщения"""
-    user_id = str(update.message.from_user.id)
+    user_id = str(update.effective_user.id)
     user_name = update.effective_user.username
 
     processing_message = await update.message.reply_text("🎧 Получил ваше голосовое сообщение. Обрабатываю аудио...")
@@ -184,16 +222,16 @@ async def handle_voice_message(update: Update, context: ContextTypes.DEFAULT_TYP
                 recognizer.adjust_for_ambient_noise(source, duration=0.25)
                 audio_data = recognizer.record(source)
 
-            # Пытаемся распознать русскую речь с помощью Google Web Speech API (бесплатно)
+            # Пытаемся распознать русскую речь с помощью Google Web Speech API
             try:
                 recognized_text = recognizer.recognize_google(audio_data, language="ru-RU")
                 logger.info(f"Распознанный текст: {recognized_text}")
 
                 # Отправляем результат пользователю
                 await update.message.reply_text(
-                    f"📝 **Распознанный текст:**\n\n"
-                    f"_{recognized_text}_",
-                    parse_mode="Markdown"
+                    f"📝 <b>Распознанный текст:</b>\n\n"
+                    f"{recognized_text}",
+                    parse_mode="HTML"
                 )
                 # Удаляем сообщение "Обрабатываю аудио..."
                 await processing_message.delete()
@@ -204,7 +242,7 @@ async def handle_voice_message(update: Update, context: ContextTypes.DEFAULT_TYP
                 # Отправляем ответ голосом!
                 await text_to_voice_and_send(update, response)
 
-                logger.info(f"Сообщение от пользователя ({user_id} - {user_name}): {response}")
+                logger.info(f"Сообщение от пользователя ({user_id} - {user_name}): {recognized_text}")
                 logger.info(f"Ответ бота: {response}\n")
 
                 coupon_message = bot.coupon_message(user_id)
@@ -232,38 +270,46 @@ async def handle_voice_message(update: Update, context: ContextTypes.DEFAULT_TYP
             await processing_message.edit_text(
                 "Произошла непредвиденная ошибка при обработке вашего голосового сообщения.")
 
+
 async def text_to_voice_and_send(update: Update, text: str):
     """Преобразует текст в голосовое сообщение и отправляет пользователю. """
     # Создаём временный файл
-    temp_file = Path(f"temp_voice_{update.message.from_user.id}.ogg")
+    temp_file = Path(f"temp_voice_{update.effective_user.id}.ogg")
 
-    try:
-        # Конвертируем текст в речь
-        tts = gTTS(text=text, lang='ru', slow=False)
+    if len(text) < 200:
+        try:
+            # Конвертируем текст в речь
+            tts = gTTS(text=text, lang='ru', slow=False)
 
-        # gTTS создаёт MP3, его нужно сохранить
-        mp3_file = temp_file.with_suffix('.mp3')
-        tts.save(str(mp3_file))
+            # gTTS создаёт MP3, его нужно сохранить
+            mp3_file = temp_file.with_suffix('.mp3')
+            tts.save(str(mp3_file))
 
-        # Конвертируем MP3 в OGG
-        from pydub import AudioSegment
-        audio = AudioSegment.from_mp3(str(mp3_file))
-        audio.export(str(temp_file), format="ogg")
+            # Конвертируем MP3 в OGG
+            audio = AudioSegment.from_mp3(str(mp3_file))
+            audio.export(str(temp_file), format="ogg")
 
-        # Отправляем голосовое сообщение
-        with open(temp_file, 'rb') as voice_file:
-            await update.message.reply_voice(voice=voice_file)
+            # Отправляем голосовое сообщение
+            with open(temp_file, 'rb') as voice_file:
+                await update.message.reply_voice(voice=voice_file)
 
-        # Удаляем временные файлы
-        if mp3_file.exists():
-            mp3_file.unlink()
-        if temp_file.exists():
-            temp_file.unlink()
+            # Удаляем временные файлы
+            if mp3_file.exists():
+                mp3_file.unlink()
+            if temp_file.exists():
+                temp_file.unlink()
 
-    except Exception as e:
-        logger.error(f"Ошибка при создании голосового сообщения: {e}")
-        # Если не получилось с голосом, отправляем текстом
-        await update.message.reply_text(text)
+        except Exception as e:
+            logger.error(f"Ошибка при создании голосового сообщения: {e}")
+            # Если не получилось с голосом, отправляем текстом
+            await update.message.reply_text(text,
+                                            reply_markup=bot.menu_keyboard,
+                                            parse_mode='HTML')
+    else:
+        await update.message.reply_text("Извиняюсь, слишком длинное сообщение для отправки голосом, отправляю в текстовом формате\n\n" + text,
+                                        reply_markup=bot.menu_keyboard,
+                                        parse_mode='HTML')
+
 
 def main():
     app = Application.builder().token(TELEGRAM_TOKEN).build()
@@ -281,7 +327,7 @@ def main():
     # Обработчик голосовых сообщений
     app.add_handler(MessageHandler(filters.VOICE, handle_voice_message))
 
-    print("Бот запущен и ожидает сообщений пользователей...")
+    logger.info("Бот запущен и ожидает сообщений пользователей...")
 
     app.run_polling()
